requirements.txt
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
pandas==1.4.0
streamlit==1.1.0
tensorflow-cpu==2.7.0
Pillow>=6.2.0
